chore/app: create ffi module and move all FFI code there
