+++ src/app/ffi/cipher_opt.rs
