+++ src/app/ffi/mutable_data/entries.rs
