+++ src/app/ffi/mutable_data/entry_actions.rs
