+++ src/app/ffi/xor_name.rs
